Updated package.json to always try last npm of react-table; Moved and componentized chainaddresses
