+++ package.json
@@ -192,7 +192,7 @@
     "react-router-dom": "^4.2.2",
     "react-select": "^1.0.0",
     "react-simple-file-input": "^2.0.0",
-    "react-table": "7.1.0",
+    "react-table": "^7.1.0",
     "react-tippy": "^1.2.2",
     "react-toggle-button": "^2.2.0",
     "react-tooltip": "^3.10.0",
